mongodb - repository migration

+++ server/src/main/scala/kpn/server/repository/NodeRouteRepositoryImpl.scala
@@ -16,39 +16,79 @@ import kpn.core.util.Log
 import org.springframework.stereotype.Component
 
 @Component
-class NodeRouteRepositoryImpl(analysisDatabase: Database) extends NodeRouteRepository {
+class NodeRouteRepositoryImpl(
+  database: Database,
+  // old
+  analysisDatabase: kpn.core.database.Database,
+  mongoEnabled: Boolean
+) extends NodeRouteRepository {
 
   private val log = Log(classOf[NodeRouteRepositoryImpl])
 
   override def save(nodeRoute: NodeRoute): Unit = {
-    log.debugElapsed {
-      analysisDatabase.save(NodeRouteDoc(docId(nodeRoute.id, nodeRoute.scopedNetworkType), nodeRoute))
-      (s"Save node-route ${nodeRoute.id}:${nodeRoute.scopedNetworkType.key}", ())
+    if (mongoEnabled) {
+      ???
+    }
+    else {
+      log.debugElapsed {
+        analysisDatabase.save(NodeRouteDoc(docId(nodeRoute.id, nodeRoute.scopedNetworkType), nodeRoute))
+        (s"Save node-route ${nodeRoute.id}:${nodeRoute.scopedNetworkType.key}", ())
+      }
     }
   }
 
   override def delete(nodeId: Long, scopedNetworkType: ScopedNetworkType): Unit = {
-    analysisDatabase.deleteDocWithId(docId(nodeId, scopedNetworkType))
+    if (mongoEnabled) {
+      ???
+    }
+    else {
+      analysisDatabase.deleteDocWithId(docId(nodeId, scopedNetworkType))
+    }
   }
 
   override def nodeRoutes(scopedNetworkType: ScopedNetworkType): Seq[NodeRoute] = {
-    NodeRouteView.query(analysisDatabase, scopedNetworkType, stale = false)
+    if (mongoEnabled) {
+      ???
+    }
+    else {
+      NodeRouteView.query(analysisDatabase, scopedNetworkType, stale = false)
+    }
   }
 
   override def nodeRouteReferences(scopedNetworkType: ScopedNetworkType, nodeId: Long): Seq[Ref] = {
-    NodeRouteReferenceView.query(analysisDatabase, scopedNetworkType, nodeId, stale = true)
+    if (mongoEnabled) {
+      ???
+    }
+    else {
+      NodeRouteReferenceView.query(analysisDatabase, scopedNetworkType, nodeId, stale = true)
+    }
   }
 
   override def nodesRouteReferences(scopedNetworkType: ScopedNetworkType, nodeIds: Seq[Long]): Seq[NodeRouteRefs] = {
-    NodeRouteReferenceView.queryNodeIds(analysisDatabase, scopedNetworkType, nodeIds, stale = true)
+    if (mongoEnabled) {
+      ???
+    }
+    else {
+      NodeRouteReferenceView.queryNodeIds(analysisDatabase, scopedNetworkType, nodeIds, stale = true)
+    }
   }
 
   override def actualNodeRouteCounts(scopedNetworkType: ScopedNetworkType): Seq[NodeRouteCount] = {
-    NodeRouteReferenceView.queryCount(analysisDatabase, scopedNetworkType, stale = false)
+    if (mongoEnabled) {
+      ???
+    }
+    else {
+      NodeRouteReferenceView.queryCount(analysisDatabase, scopedNetworkType, stale = false)
+    }
   }
 
   override def expectedNodeRouteCounts(scopedNetworkType: ScopedNetworkType): Seq[NodeRouteExpectedCount] = {
-    NodeRouteExpectedView.queryScopedNetworkType(analysisDatabase, scopedNetworkType, stale = false)
+    if (mongoEnabled) {
+      ???
+    }
+    else {
+      NodeRouteExpectedView.queryScopedNetworkType(analysisDatabase, scopedNetworkType, stale = false)
+    }
   }
 
   private def docId(nodeId: Long, scopedNetworkType: ScopedNetworkType): String = {
